Move decorator utils into `decorators`
